further cleaning & moving low-level functions to fluxcal module
